% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{check_load_data_button}
\alias{check_load_data_button}
\title{Evaluates whether the load-data button needs to be shown or not.}
\usage{
check_load_data_button(rv, session)
}
\arguments{
\item{rv}{The global 'reactiveValues()' object, defined in server.R}

\item{session}{Shiny session object}
}
\description{
If there is a valid source system and a valid target system
(this is also the case if the user set target == source), the result
of this function will be TRUE and the button will be displayed via
shinyjs. Otherwise the result is FALSE and the button will be hidden.
This function also displays (or hides) the variables which can be
assessed.
}
