% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{feedback_txt}
\alias{feedback_txt}
\title{This function is used in the config-tab and displays the selected
system to the user.}
\usage{
feedback_txt(system, type)
}
\arguments{
\item{system}{(String) e.g. "i2b2", "OMOP" or "CSV"}

\item{type}{(String) "source" or "target"}
}
\value{
String containing the input params in a propper manner
}
\description{
This function is used in the config-tab and displays the selected
system to the user.
}
