% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{get_db_settings}
\alias{get_db_settings}
\title{get_db_settings}
\usage{
get_db_settings(input, target = TRUE, db_type, displayname_gui, rv)
}
\arguments{
\item{input}{Shiny server input object.}

\item{target}{A boolean (default: TRUE).}

\item{db_type}{(String) "postgres" or "oracle".}

\item{displayname_gui}{(String) "i2b2 (Prod)"}

\item{rv}{The global 'reactiveValues()' object, defined in server.R}
}
\value{
This functions returns a data table of key-value pairs for the
database settings, which are parsed from the respective config tab
from the shiny application.
}
\description{
get_db_settings
}
\examples{
if (interactive()) {
 get_db_settings(
  input = input,
  target = TRUE,
  db_type = "postgres"
 )
}

}
