% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleAtempPlausibility.R
\name{module_atemp_pl_ui}
\alias{module_atemp_pl_ui}
\title{module_atemp_pl_ui}
\usage{
module_atemp_pl_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\value{
The function returns a shiny ui module.
}
\description{
module_atemp_pl_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "atemp_plausi",
    module_atemp_pl_ui(
      "moduleAtemporalPlausibilities"
    )
  )
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
