% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleCompleteness.R
\name{module_completeness_ui}
\alias{module_completeness_ui}
\title{module_completeness_ui}
\usage{
module_completeness_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\value{
The function returns a shiny ui module.
}
\description{
module_completeness_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "completeness",
    module_completeness_ui(
      "moduleCompleteness"
    )
  )
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
