% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleDescriptive.R
\name{module_descriptive_ui}
\alias{module_descriptive_ui}
\title{module_descriptive_ui}
\usage{
module_descriptive_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\value{
The function returns a shiny ui module.
}
\description{
module_descriptive_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "descriptive",
    module_descriptive_ui(
      "moduleDescriptive"
    )
  )
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
