% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleDifferences.R
\name{module_differences_ui}
\alias{module_differences_ui}
\title{module_differences_ui}
\usage{
module_differences_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\value{
The function returns a shiny ui module.
}
\description{
module_differences_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "differences",
    module_differences_ui(
      "moduleDifferences"
    )
  )
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
