% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleLog.R
\name{module_log_ui}
\alias{module_log_ui}
\title{module_log_ui}
\usage{
module_log_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\value{
The function returns a shiny ui module.
}
\description{
module_log_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "logging",
    module_log_ui(
      "moduleLogging"
    )
  )
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
