% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleMDR.R
\name{module_mdr_ui}
\alias{module_mdr_ui}
\title{module_mdr_ui}
\usage{
module_mdr_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\value{
The function returns a shiny ui module.
}
\description{
module_mdr_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "mdr",
    module_mdr_ui(
      "moduleMDR"
    )
  )
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
