% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleReport.R
\name{module_report_server}
\alias{module_report_server}
\title{module_report_server}
\usage{
module_report_server(input, output, session, rv, input_re)
}
\arguments{
\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny session object}

\item{rv}{The global 'reactiveValues()' object, defined in server.R}

\item{input_re}{The Shiny server input object, wrapped into a reactive
expression: \code{input_re = reactive({input})}}
}
\value{
The function returns a shiny server module.
}
\description{
module_report_server
}
\examples{
if (interactive()) {
rv <- list()
shiny::callModule(
  module_report_server,
  "moduleReport",
  rv = rv,
  input_re = reactive(input)
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
