% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleReport.R
\name{module_report_ui}
\alias{module_report_ui}
\title{module_report_ui}
\usage{
module_report_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\value{
The function returns a shiny ui module.
}
\description{
module_report_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "report",
    module_report_ui(
      "moduleReport"
    )
  )
)
}
}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
