% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleUniquePlausibility.R
\name{module_uniq_plaus_ui}
\alias{module_uniq_plaus_ui}
\title{module_uniq_plaus_ui}
\usage{
module_uniq_plaus_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\value{
The function returns a shiny ui module.
}
\description{
module_uniq_plaus_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "uniq_plausis",
    module_uniq_plaus_ui(
      "moduleUniquenessPlausibilities"
    )
  )
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
