% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{set_target_equal_to_source}
\alias{set_target_equal_to_source}
\title{This function is called when the user clicks on the button}
\usage{
set_target_equal_to_source(rv)
}
\arguments{
\item{rv}{The global 'reactiveValues()' object, defined in server.R}
}
\description{
"Set target == source". It sets target settings = source
settings.
}
