% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{show_failure_alert}
\alias{show_failure_alert}
\title{Sjows an error alert modal with the hint to look into the logfile.}
\usage{
show_failure_alert(what_failed)
}
\arguments{
\item{what_failed}{Short description of what failed.Like:
"Getting the data failed."
'}
}
\value{
Nothing - Just shows the alert modal.
}
\description{
See title.
}
