% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{validate_inputs}
\alias{validate_inputs}
\title{This function checks if all necessary input parameters
  for source and target exist and are valid.}
\usage{
validate_inputs(rv, input, output, session)
}
\arguments{
\item{rv}{The global 'reactiveValues()' object, defined in server.R}

\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny session object}
}
\description{
This function checks if all necessary input parameters
  for source and target exist and are valid.
}
